requirements.txt
# requirements.txt
fastapi
uvicorn[standard]
python-dotenv
sqlalchemy
pymysql
langchain-openai
langchain-community
langchain-google-community
google-api-python-client
pydantic
pandas
playwright
chromadb